Character API > Folder refactoring
