+++ lib/Character_Wiki/Data/Character/character_model.dart
