+++ lib/Character_Wiki/Data/Character/character_remote_data_source.dart
@@ -1,7 +1,7 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'dart:convert';
 
-import 'package:character_wiki/Character_Wiki/Data/character_model.dart';
+import 'package:character_wiki/Character_Wiki/Data/Character/character_model.dart';
 import 'package:http/http.dart' as http;
 
 class CharacterRemoteDataSource {
