+++ lib/Character_Wiki/Domain/Character/character.dart
