+++ lib/Character_Wiki/Domain/Character/character_repository_impl.dart
@@ -1,7 +1,7 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
-import 'package:character_wiki/Character_Wiki/Data/character_remote_data_source.dart';
-import 'package:character_wiki/Character_Wiki/Domain/character.dart';
-import 'package:character_wiki/Character_Wiki/Domain/character_repository.dart';
+import 'package:character_wiki/Character_Wiki/Data/Character/character_remote_data_source.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Character/character.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Character/character_repository.dart';
 
 class CharacterRepositoryImpl implements CharacterRepository {
   final CharacterRemoteDataSource remoteDataSource;
