+++ lib/Character_Wiki/Domain/Character/get_characters.dart
@@ -1,6 +1,6 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
-import 'package:character_wiki/Character_Wiki/Domain/character.dart';
-import 'package:character_wiki/Character_Wiki/Domain/character_repository.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Character/character.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Character/character_repository.dart';
 
 class GetCharacters {
   final CharacterRepository repository;
