+++ lib/Character_Wiki/Presentation/Bloc/Character/character_event.dart
