+++ lib/Character_Wiki/Presentation/Bloc/Character/character_state.dart
