+++ lib/service_locator.dart
@@ -1,8 +1,8 @@
-import 'package:character_wiki/Character_Wiki/Data/character_remote_data_source.dart';
-import 'package:character_wiki/Character_Wiki/Domain/character_repository.dart';
-import 'package:character_wiki/Character_Wiki/Domain/character_repository_impl.dart';
-import 'package:character_wiki/Character_Wiki/Domain/get_characters.dart';
-import 'package:character_wiki/Character_Wiki/Presentation/Bloc/character_bloc.dart';
+import 'package:character_wiki/Character_Wiki/Data/Character/character_remote_data_source.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Character/character_repository.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Character/character_repository_impl.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Character/get_characters.dart';
+import 'package:character_wiki/Character_Wiki/Presentation/Bloc/Character/character_bloc.dart';
 import 'package:get_it/get_it.dart';
 import 'package:http/http.dart' as http;
